+++ CHANGELOG.md
@@ -23,6 +23,7 @@ any pages related to that taxonomy
 - Use Zola Tera instance for markdown filter: this means you have access to the same Tera functions as in shortcodes
 - Ignore sections with `render=false` when looking for path collisions
 - Add support for backlinks
+- Add a warning mode for internal/external link checking in case you don't want zola to stop the build on invalid links
 
 ## 0.15.3 (2022-01-23)